Split BindingExtensions class
