+++ sources/Google.Solutions.IapDesktop.Application/Views/Options/GeneralOptionsSheet.cs
@@ -21,6 +21,7 @@
 
 using Google.Solutions.Common.Diagnostics;
 using Google.Solutions.Mvvm.Binding;
+using Google.Solutions.Mvvm.Binding.Commands;
 using System;
 using System.Windows.Forms;
