+++ sources/Google.Solutions.Mvvm/Binding/Commands/CommandContainer.cs
@@ -20,7 +20,6 @@
 //
 
 using Google.Solutions.Common.Util;
-using Google.Solutions.Mvvm.Binding;
 using Google.Solutions.Mvvm.Controls;
 using System;
 using System.Collections.Generic;
